modified:   todo/src/main/java/com/todo/todo/controller/todoController.java 	modified:   todo/src/main/java/com/todo/todo/entity/Todo.java 	modified:   todo/src/main/java/com/todo/todo/repository/TodoRepository.java 	modified:   todo/src/main/java/com/todo/todo/service/TodoService.java 	modified:   todo/target/classes/com/todo/todo/controller/todoController.class 	modified:   todo/target/classes/com/todo/todo/entity/Todo.class 	modified:   todo/target/classes/com/todo/todo/repository/TodoRepository.class 	modified:   todo/target/classes/com/todo/todo/service/TodoService.class 	modified:   todo/target/test-classes/com/todo/todo/TodoApplicationTests.class

+++ todo/src/main/java/com/todo/todo/repository/TodoRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface TodoRepository extends JpaRepository<Todo, Long> {
+public interface TodoRepository extends JpaRepository<Todo, Integer> {
     List<Todo> findByTitle(String title);
 }
